refactor: decompose fileOperations
