+++ file-manager/main.js
@@ -7,7 +7,7 @@ import { closeSession, getCurrentlyPath, getUsername } from './utils/helpers.js'
 import { parseInput } from './utils/inputParser.js';
 import { cd, up } from './commands/navigation.js';
 import { ls } from './commands/fileManager.js';
-import { add, cat, cp, rm, rn } from './commands/fileOperations.js';
+import { add, cat, cp, mv, rm, rn } from './commands/fileOperations.js';
 
 const rl = createInterface({
     input: stdin,
@@ -52,6 +52,10 @@ rl.on('line', async (line) => {
             }
             case 'cp': {
                 await cp(args[0], args[1]);
+                break;
+            }
+            case 'mv': {
+                await mv(args[0], args[1]);
                 break;
             }
             case '.exit': {
